SpotifyJS part 3 - fix: corresponding song playing on click on div
