+++ spotify/index.html
@@ -7,6 +7,7 @@
     <link rel="stylesheet" href="css/styles.css" />
     <link rel="stylesheet" href="css/fonts.css" />
     <link rel="stylesheet" href="css/media_screen.css" />
+    <script src="js/main.js" type="module"></script>
     <script
       src="https://kit.fontawesome.com/d03b1f5ced.js"
       crossorigin="anonymous"
@@ -265,6 +266,5 @@
         </div>
       </div>
     </div>
-    <script src="js/main.js"></script>
   </body>
 </html>
